⬆️ 升级到 docusaurus v3
