+++ babel.config.js
@@ -1,3 +1,3 @@
 module.exports = {
-  presets: [require.resolve('@docusaurus/core/lib/babel/preset'), ['@babel/preset-react', { runtime: 'automatic' }]],
+  presets: [require.resolve('@docusaurus/core/lib/babel/preset')],
 }
