+++ docs/guide/start/installation.mdx
@@ -4,7 +4,7 @@ title: 安装
 
 在本章节中，我们将使用 Kirami CLI 脚手架创建和运行你的第一个 KiramiBot 项目。
 
-:::caution 注意
+:::warning 注意
 在开始安装 KiramiBot 之前，请确保满足以下条件：
 
 - Python 版本 >= 3.10
