+++ docs/guide/tutorial/plugin.mdx
@@ -6,7 +6,7 @@ title: 插件
 
 ## KiramiBot 插件
 
-:::caution 注意
+:::warning 注意
 KiramiBot 的商店还在建设中，目前只有少量插件可用。如果你有好的插件想要分享，请前往 [KiramiBot 插件目录](https://github.com/A-kirami/kirami-plugins-index) 提交 PR。
 :::
