+++ docs/manual/depth/api-calling.mdx
@@ -46,6 +46,6 @@ result = await bot.get_user_info(user_id=12345678)
 result = await bot.call_api("get_user_info", user_id=12345678)
 ```
 
-:::caution 注意
+:::warning 注意
 实际可以使用的 API 以及参数取决于平台提供的接口以及协议适配器的实现，KiramiBot 使用 OneBotV11 协议，支持的 API 可以参考 [OneBotV11 API](https://github.com/botuniverse/onebot-11/blob/master/api/public.md)。
 :::
