+++ docs/manual/depth/scheduler.mdx
@@ -16,7 +16,7 @@ async def run_every_10_seconds():
     pass
 ```
 
-:::caution 注意
+:::warning 注意
 `on_time` 无法使用 `Matcher` 对象直接发送消息，因为它是由 Bot 自身事件触发，而不是外部事件。我们需要通过[调用平台 API](api-calling#调用平台-api)的方式来获取信息或收发消息。
 :::
 
@@ -41,7 +41,7 @@ scheduler.add_job(
 )
 ```
 
-:::warning 注意
+:::danger 注意
 由于 APScheduler 的定时任务并不是**由事件响应器所触发的事件**，因此其任务函数无法同[事件处理函数](../basic/handler#事件处理函数)一样通过[依赖注入](../basic/event-data#初识依赖注入)获取上下文信息，也无法通过事件响应器对象的方法进行任何操作，因此我们需要通过[调用平台 API](api-calling#调用平台-api)的方式来获取信息或收发消息。
 
 相对于事件处理依赖而言，编写定时任务更像是编写普通的函数，需要我们自行获取信息以及发送信息，请**不要**将事件处理依赖的特殊语法用于定时任务！
