+++ docs/manual/expand/request.mdx
@@ -1,3 +1,3 @@
 ---
-title: 发送请求
+title: 网络请求
 ---
