+++ docs/manual/index.mdx
@@ -11,7 +11,7 @@ title: 准备工作
 
 </div>
 
-:::caution 注意
+:::warning 注意
 在官方开发手册中，我们假设你已掌握了 [Python](https://docs.python.org/zh-cn/) 的初级知识，并且对 [asyncio](https://docs.python.org/zh-cn/3/library/asyncio.html) 有基本的了解。
 如果你刚开始学习 Python，将 KiramiBot 插件开发作为你的第一步并不是一个好主意，掌握好基础知识再来吧！
 :::
