+++ package.json
@@ -20,46 +20,46 @@
     "postinstall": "husky install"
   },
   "dependencies": {
-    "@docusaurus/core": "^2.4.1",
-    "@docusaurus/preset-classic": "^2.4.1",
-    "@markprompt/docusaurus-theme-search": "^0.9.0",
-    "@mdx-js/react": "^1.6.22",
-    "clsx": "^1.2.1",
-    "prism-react-renderer": "^1.3.5",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2"
+    "@docusaurus/core": "^3.0.1",
+    "@docusaurus/preset-classic": "^3.0.1",
+    "@markprompt/docusaurus-theme-search": "^0.13.15",
+    "@mdx-js/react": "^3.0.0",
+    "clsx": "^2.0.0",
+    "prism-react-renderer": "^2.3.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0"
   },
   "devDependencies": {
-    "@docusaurus/module-type-aliases": "^2.4.1",
-    "@iconify-json/ri": "^1.1.12",
-    "@tsconfig/docusaurus": "^1.0.7",
-    "@typescript-eslint/eslint-plugin": "^6.3.0",
-    "@typescript-eslint/parser": "^6.3.0",
-    "@unocss/eslint-config": "^0.55.0",
-    "@unocss/postcss": "^0.55.0",
-    "@unocss/preset-icons": "^0.55.0",
-    "esbuild": "^0.19.2",
-    "eslint": "^8.47.0",
-    "eslint-config-prettier": "^9.0.0",
-    "eslint-import-resolver-typescript": "^3.6.0",
-    "eslint-plugin-import": "^2.28.0",
-    "eslint-plugin-mdx": "^2.2.0",
-    "eslint-plugin-prettier": "^5.0.0",
+    "@docusaurus/module-type-aliases": "^3.0.1",
+    "@docusaurus/tsconfig": "^3.0.1",
+    "@docusaurus/types": "^3.0.1",
+    "@iconify-json/ri": "^1.1.17",
+    "@typescript-eslint/eslint-plugin": "^6.14.0",
+    "@typescript-eslint/parser": "^6.14.0",
+    "@unocss/eslint-config": "^0.58.0",
+    "@unocss/postcss": "^0.58.0",
+    "esbuild": "^0.19.9",
+    "eslint": "^8.56.0",
+    "eslint-config-prettier": "^9.1.0",
+    "eslint-import-resolver-typescript": "^3.6.1",
+    "eslint-plugin-import": "^2.29.1",
+    "eslint-plugin-mdx": "^3.0.0",
+    "eslint-plugin-prettier": "^5.0.1",
     "eslint-plugin-promise": "^6.1.1",
-    "eslint-plugin-react": "^7.33.1",
+    "eslint-plugin-react": "^7.33.2",
     "husky": "^8.0.3",
-    "lint-staged": "^13.2.3",
-    "postcss-nesting": "^12.0.1",
-    "prettier": "^3.0.1",
-    "stylelint": "^15.10.2",
+    "lint-staged": "^15.2.0",
+    "postcss-nesting": "^12.0.2",
+    "prettier": "^3.1.1",
+    "stylelint": "^16.0.2",
     "stylelint-color-format": "^1.1.0",
     "stylelint-config-css-modules": "^4.3.0",
-    "stylelint-config-recess-order": "^4.3.0",
-    "stylelint-config-standard": "^34.0.0",
-    "stylelint-config-standard-scss": "^10.0.0",
+    "stylelint-config-recess-order": "^4.4.0",
+    "stylelint-config-standard": "^35.0.0",
+    "stylelint-config-standard-scss": "^12.0.0",
     "stylelint-declaration-block-no-ignored-properties": "^2.7.0",
-    "typescript": "^4.9.5",
-    "unocss": "^0.55.0"
+    "typescript": "^5.3.3",
+    "unocss": "^0.58.0"
   },
   "packageManager": "yarn@3.6.1",
   "browserslist": {
@@ -69,13 +69,13 @@
       "not op_mini all"
     ],
     "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
+      "last 3 chrome version",
+      "last 3 firefox version",
+      "last 5 safari version"
     ]
   },
   "engines": {
-    "node": ">=18.13"
+    "node": ">=18.0"
   },
   "husky": {
     "hooks": {
