+++ src/theme/DocItem/Layout/index.tsx
@@ -8,6 +8,7 @@ import DocItemTOCDesktop from '@theme/DocItem/TOC/Desktop'
 import DocItemTOCMobile from '@theme/DocItem/TOC/Mobile'
 import DocVersionBadge from '@theme/DocVersionBadge'
 import DocVersionBanner from '@theme/DocVersionBanner'
+import Unlisted from '@theme/Unlisted'
 import clsx from 'clsx'
 
 import styles from './styles.module.css'
@@ -37,9 +38,13 @@ function useDocTOC() {
 
 export default function DocItemLayout({ children }: Props): JSX.Element {
   const docTOC = useDocTOC()
+  const {
+    metadata: { unlisted },
+  } = useDoc()
   return (
     <div className="row">
       <div className={clsx('col', !docTOC.hidden && styles.docItemCol)}>
+        {unlisted && <Unlisted />}
         <DocVersionBanner />
         <div className={styles.docItemContainer}>
           <article>
