+++ src/theme/NotFound/index.tsx
@@ -1,0 +1,19 @@
+import { PageMetadata } from '@docusaurus/theme-common'
+import { translate } from '@docusaurus/Translate'
+import Layout from '@theme/Layout'
+import NotFoundContent from '@theme/NotFound/Content'
+
+export default function Index(): JSX.Element {
+  const title = translate({
+    id: 'theme.NotFound.title',
+    message: 'Page Not Found',
+  })
+  return (
+    <>
+      <PageMetadata title={title} />
+      <Layout>
+        <NotFoundContent />
+      </Layout>
+    </>
+  )
+}
